requirements-dev.txt
pandas>=2.0
numpy>=1.24
google-cloud-bigquery>=3.0
pyarrow>=14.0
